mv histogram calc2d to histogram2d/
